refactor: move page-list component to page folder
