--- example/src/ui/layout/page-list/index.ts
@@ -1,1 +1,0 @@
-export { PageList, type PageProps } from './page-list';
