--- example/src/ui/layout/page-list/page-list-footer.tsx
@@ -1,4 +1,0 @@
-export {
-  PageListFooter,
-  type PageListFooterProps,
-} from '@kode-frontend/uikit-rn/layouts';
