+++ example/src/ui/layout/page/page/index.ts
@@ -1,0 +1,1 @@
+export { Page, type PageProps } from './page';
