--- src/layouts/page-list/index.ts
@@ -1,2 +1,0 @@
-export { PageList, type PageListProps } from './page-list';
-export { PageListFooter, type PageListFooterProps } from './page-list-footer';
