+++ src/layouts/page/index.ts
@@ -1,2 +1,3 @@
 export { Page, type PageProps } from './page';
+export { PageList, type PageListProps } from './page-list';
 export { PageFooter, type PageFooterProps } from './page-footer';
